Renaming Source/mock files to lhc_mock
