+++ Source/Mock/lhc_mock.cpp
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 #include "pch.h"
-#include "mock.h"
+#include "lhc_mock.h"
 #include "../HTTP/httpcall.h"
 
 using namespace xbox::httpclient;
